+++ apis/vara-eth/src/types/api/index.ts
@@ -1,3 +1,5 @@
 export * from './block.js';
+export * from './common.js';
+export * from './message.js';
 export * from './program.js';
 export * from './injected.js';